ITB-85: refactor skeleton for tests
