+++ src/test/java/com/inthebytes/restaurantmanager/control/RestaurantControllerTest.java
@@ -1,0 +1,5 @@
+package com.inthebytes.restaurantmanager.control;
+
+public class RestaurantControllerTest {
+
+}
