+++ src/test/java/com/inthebytes/restaurantmanager/dto/RestaurantDtoTest.java
@@ -1,0 +1,5 @@
+package com.inthebytes.restaurantmanager.dto;
+
+public class RestaurantDtoTest {
+
+}
