+++ src/test/java/com/inthebytes/restaurantmanager/service/RestaurantServiceTest.java
@@ -1,0 +1,5 @@
+package com.inthebytes.restaurantmanager.service;
+
+public class RestaurantServiceTest {
+
+}
